Recognition/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dlib==19.24.4
dnspython==2.6.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.3
flatbuffers==24.3.25
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
keras==2.15.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.7
numpy==1.26.4
oauthlib==3.2.2
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.3
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.13.0
six==1.16.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.15.0
termcolor==2.4.0
threadpoolctl==3.4.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.14.1
zipp==3.18.1
